requirements.txt
absl-py==1.0.0
anyio==3.4.0
asgiref==3.4.1
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cycler==0.11.0
fastapi==0.70.1
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
h11==0.12.0
h5py==3.6.0
idna==3.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-serving-api==2.5.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
Werkzeug==2.0.3
wrapt==1.13.3
